Fixed casting of order to int.

+++ hcipy/coronagraphy/perfect_coronagraph.py
@@ -12,9 +12,9 @@ class PerfectCoronagraph(OpticalElement):
 			order = 2 * np.ceil(0.5 * (np.sqrt(8*len(coeffs) + 1) - 1))
 			self.coeffs = coeffs
 		else:
-			self.coeffs = np.ones(order * (order / 2 + 1) / 4)
+			self.coeffs = np.ones(int(order * (order / 2 + 1) / 4))
 
-		for i in range(order / 2):
+		for i in range(int(order / 2)):
 			for j in range(i + 1):
 				modes.append(aperture * self.pupil_grid.x**j * self.pupil_grid.y**(i-j))
